refactor(server): Type API handlers and refactor error class - pokemonHandler is refactored to a generic function. - Custom API error classes and types moved to server\types\. - Implement a typed error response.
