+++ server/api/pokemon/[name]/index.get.ts
@@ -1,7 +1,8 @@
 import pokemonHandler from "@@/server/utils/pokemon-handler";
 import createPokemonClient from "@@/server/client/pokemon-worker";
+import type { PokemonProfile} from "@@/server/types/pokemon-api"
 
 export default defineEventHandler(async (event) => {
   const client = createPokemonClient(event.context.cloudflare.env.pokemon);
-  return await pokemonHandler(event, client.getPokemon);
+  return await pokemonHandler<PokemonProfile>(event, client.getPokemon);
 });
