+++ server/api/pokemon/[name]/locations.get.ts
@@ -1,7 +1,8 @@
 import pokemonHandler from "@@/server/utils/pokemon-handler";
 import createPokemonClient from "@@/server/client/pokemon-worker";
+import type { PokemonSpawnLocation } from "@@/server/types/pokemon-api";
 
 export default defineEventHandler(async (event) => {
   const client = createPokemonClient(event.context.cloudflare.env.pokemon);
-  return await pokemonHandler(event, client.getPokemonLocations);
+  return await pokemonHandler<PokemonSpawnLocation>(event, client.getPokemonLocations);
 });
