+++ server/api/pokemon/[name]/moves.get.ts
@@ -1,8 +1,8 @@
 import pokemonHandler from "@@/server/utils/pokemon-handler";
 import createPokemonClient from "@@/server/client/pokemon-worker";
+import type { PokemonMoves } from "@@/server/types/pokemon-api";
 
 export default defineEventHandler(async (event) => {
   const client = createPokemonClient(event.context.cloudflare.env.pokemon);
-
-  return await pokemonHandler(event, client.getPokemonMoves);
+  return await pokemonHandler<PokemonMoves>(event, client.getPokemonMoves);
 });
